// frontend/src/client/types.gen.ts
export type MediaType = 'text' | 'voice' | 'image' | 'link' | 'pdf';

/**
 * NoteCreate
 */
export type NoteCreate = {
    /**
     * Content
     */
    content: string;
    media_type?: MediaType;
    /**
     * Tags
     */
    tags?: Array<string>;
    /**
     * File Path
     */
    file_path?: string | null;
    /**
     * Parent Id
     */
    parent_id?: number | null;
    /**
     * Is Hidden
     */
    is_hidden?: boolean;
};

/**
 * NoteResponse
 */
export type NoteResponse = {
    /**
     * Content
     */
    content: string;
    media_type?: MediaType;
    /**
     * Tags
     */
    tags?: Array<string>;
    /**
     * Id
     */
    id: number;
    /**
     * Summary
     */
    summary?: string | null;
    /**
     * File Path
     */
    file_path?: string | null;
    /**
     * Parent Id
     */
    parent_id?: number | null;
    /**
     * Category
     */
    category?: string | null;
    /**
     * Event At
     */
    event_at?: string | null;
    /**
     * Event Duration
     */
    event_duration?: number;
    /**
     * Is Hidden
     */
    is_hidden?: boolean;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
    /**
     * Is Active
     */
    is_active: boolean;
    /**
     * Is Processing
     */
    is_processing?: boolean;
};